requirements.txt
Django===2.1.1
markdown===2.6.11
django-filter===2.0.0
djangorestframework===3.8.2
dnspython===1.15.0
gevent===1.3.6
greenlet===0.4.14
esd===0.0.11
mysqlclient===1.3.13